Move campaign & its logic from "search result list" to "search result"
This makes sense as the campaign isn't a part of the search result list,
but rather just a different component that shows on the page under
specific circumstances.
